// src/app/models/marriage.ts
import { Feature } from "./missing";

export class Marriage {
  groomFName: string;
  brideFName: string;
  groomLName: string;
  brideLName: string;
  mainType: string;
  mainCertificateNo:string;
  status:string;
  date: string;
  time: string;
  image: string;
  toast:string;
  state:string;
  city:string;
  postalCode:string;
  createdBy:string;
  countryID: string;
  weddingDate:string;
  id:string;
  dateCreated:string;
}